requirements.txt
flask
torch
torchvision
pillow
faiss-cpu
numpy
pandas
